+++ src/Container/FreezingContainer.php
@@ -6,14 +6,13 @@ use Krak\Cargo;
 
 /** Freezes a service definition after it's been used preventing an override of that
     service */
-class FreezingContainer extends AbstractContainer
+class FreezingContainer extends ContainerDecorator
 {
-    private $container;
     private $only_services;
     private $frozen;
 
     public function __construct(Cargo\Container $container, $only_services = true) {
-        $this->container = $container;
+        parent::__construct($container);
         $this->only_services = $only_services;
         $this->frozen = [];
     }
@@ -28,23 +27,11 @@ class FreezingContainer extends AbstractContainer
 
         return $this->container->get($id, $container ?: $this);
     }
-    public function has($id) {
-        return $this->container->has($id);
-    }
-    public function remove($id) {
-        return $this->container->remove($id);
-    }
     public function add($id, $box, array $opts = []) {
         if (array_key_exists($id, $this->frozen)) {
             throw new Cargo\Exception\ContainerException("Cannot redefine already frozen service '$id'");
         }
 
         return $this->container->add($id, $box, $opts);
-    }
-    public function box($id) {
-        return $this->container->box($id);
-    }
-    public function keys() {
-        return $this->container->keys();
     }
 }